+++ kernels/bvh/bvh_rotate.cpp
@@ -40,7 +40,7 @@ namespace embree
 	(child2child) of a different second child (child2), and swap child1 
 	and child2child. We perform the best such swap. */
       float bestArea = 0;
-      size_t bestChild1 = -1, bestChild2 = -1, bestChild2Child = -1;
+      size_t bestChild1 = (size_t)-1, bestChild2 = (size_t)-1, bestChild2Child = (size_t)-1;
       for (size_t c2=0; c2<4; c2++)
       {
 	/*! ignore leaf nodes as we cannot descent into them */
